backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pdfplumber
google-genai
numpy
python-multipart
python-dotenv
pydantic
aiofiles
